requirements.txt
elasticsearch>=1.3.0,<2.0 # Apache-2.0
falcon>=0.1.6 # Apache-2.0
jsonschema>=2.0.0,<3.0.0,!=2.5.0 # MIT
keystonemiddleware>=4.0.0 # Apache-2.0
Paste  # MIT
PasteDeploy>=1.5.0  # MIT
oslo.config>=3.2.0 # Apache-2.0
oslo.context>=2.2.0                     # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.log>=1.14.0  # Apache-2.0
six>=1.9.0 # MIT
python-cinderclient # Apache-2.0
python-swiftclient # Apache-2.0
python-glanceclient # Apache-2.0
python-novaclient # Apache-2.0
keystoneauth1 # Apache-2.0
